+++ CHANGELOG.md
@@ -79,6 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix `--pretty` option of the `export` command having no effect on any output format
 - Fix `explain` command rejecting structure types given by their number
 - Fix name of string property value attribute
 - Fix false errors on memory device references to handle 0x0000 used as unspecified value